cleaned project
