+++ components/CarouselImage.js
@@ -1,19 +1,19 @@
 /* eslint-disable @next/next/no-img-element */
 import { shrinkHash } from "@/lib/utils";
-export default function CarouselImage({ hash }) {
+
+export default function CarouselImage({ contractAddress, tokenId, img, big }) {
   return (
     <>
-      <div className="lg:w-max lg:flex-shrink-0 text-gray-600 hover:text-gray-300 transition-colors">
-        <div className="flex-col ">
-          <a href={`/pieces/${hash}`}>
+      <div className="text-gray-600 hover:text-gray-300 transition-colors mx-auto">
+        <div className={"flex-col "}>
+          <a href={`/collections/${contractAddress}/${tokenId}`}>
             <img
-              src={`https://${hash}.ipfs.infura-ipfs.io/`}
+              src={img}
               alt=""
-              width={500}
-              height={500}
+              className="w-[180px] h-[180px] md:w-[230px] md:h-[230px] xl:w-[180px] xl:h-[180px]"
             />
           </a>
-          <div className="w-48 font-body truncate ...">{shrinkHash(hash)}</div>
+          <div className="font-body truncate ...">{shrinkHash("hash")}</div>
         </div>
       </div>
     </>
