--- components/hashes.js
@@ -1,5 +1,0 @@
-export var hashes = [
-  "bafybeidwzynbvhtc6qspgkn7amkgd4nacubpxoezul4qroiqdf2vympr3u",
-  "bafybeiaobfbg56q6vcmcs7o3qya4enftunly36akfzei4oyznwv4633o54",
-  "bafybeiarr6v3ajdv5iwiwmehmcn4tqxpdpnenuokrgvon5odg6ujev2xvi",
-];
