+++ package.json
@@ -12,7 +12,6 @@
     "@pinata/sdk": "^1.1.25",
     "@radix-ui/react-icons": "^1.1.1",
     "@usedapp/core": "^1.0.2",
-    "axios": "^0.27.2",
     "form-data": "^4.0.0",
     "formidable": "^2.0.1",
     "fs": "^0.0.1-security",
